backend/requirements.txt
fastapi
uvicorn
gunicorn
pymongo
scikit-learn
pandas
pydantic
pydantic-settings
python-dotenv
python-multipart
motor
